remove unnecessary engineering for now

+++ src/main/java/com/rodrigocso/groceries/repository/BrandRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface BrandRepository extends JpaRepository<Brand, Integer> {
+public interface BrandRepository extends JpaRepository<Brand, Long> {
     Optional<Brand> findByNameIgnoreCase(String name);
     List<Brand> findByNameContainingIgnoreCase(String partialName);
 }
