+++ src/main/java/com/rodrigocso/groceries/repository/ItemRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
-    List<Item> findByProductId(Integer productId);
+public interface ItemRepository extends JpaRepository<Item, Long> {
+    List<Item> findByProductId(Long productId);
 }
