+++ src/main/java/com/rodrigocso/groceries/repository/ProductRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-    List<Product> findByBrandId(Integer brandId);
-    Optional<Product> findByBrandIdAndName(Integer brandId, String name);
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    List<Product> findByBrandId(Long brandId);
+    Optional<Product> findByBrandIdAndName(Long brandId, String name);
     List<Product> findByNameContainingIgnoreCase(String name);
 }
